UI Cleanuo

+++ src/Babana/Models/PathTraceSnapshot.cs
@@ -6,9 +6,9 @@ public class PathTraceSnapshot {
     public PathTraceSnapshot() {
     }
 
-    public double AveRespTime { get; set; }
-    public double Throughput { get; set; }
-    public double P90RespTime { get; set; }
+    public float AveRespTime { get; set; }
+    public float Throughput { get; set; }
+    public float P90RespTime { get; set; }
     public string Path { get; set; }
 
     public string Host { get; set; }
