+++ src/Babana/Models/PathTraceSnapshotItem.cs
@@ -1,8 +1,8 @@
 namespace PlaywrightTest.Models;
 
 public class PathTraceSnapshotItem {
-    public double AveRespTime { get; set; }
-    public double Throughput { get; set; }
-    public double P90RespTime { get; set; }
+    public float AveRespTime { get; set; }
+    public float Throughput { get; set; }
+    public float P90RespTime { get; set; }
     public int VirtualUserCount { get; set; }
 }
